move poll creation page to /
